+++ src/constants.js
@@ -4,7 +4,7 @@
  */
 
 // Application Version
-export const APP_VERSION = '1.4.1';
+export const APP_VERSION = '1.5.0';
 
 // Theme Colors
 export const THEME_COLORS = {